Fix: Rename file to match component
