+++ ui/src/app/components/gke-app-configuration-panel.spec.tsx
@@ -12,8 +12,8 @@ import {
 } from 'app/components/gke-app-configuration-panel';
 import { RStudioConfigurationPanel } from 'app/components/rstudio-configuration-panel';
 import { DEFAULT_PROPS as RSTUDIO_DEFAULT_PROPS } from 'app/components/rstudio-configuration-panel.spec';
+import { ConfirmDeleteEnvironmentWithPD } from 'app/components/runtime-configuration-panel/confirm-delete-environment-with-pd';
 import { ConfirmDeleteUnattachedPD } from 'app/components/runtime-configuration-panel/confirm-delete-unattached-pd';
-import { ConfirmDeleteEnvironmentWithPD } from 'app/components/runtime-configuration-panel/confirm-environment-delete-with-pd';
 import { Spinner } from 'app/components/spinners';
 import {
   appsApi,
