+++ ui/src/app/components/runtime-configuration-panel/confirm-delete-environment-with-pd.tsx
